Refactor source tree into f3 and f3-extra
