+++ include/f3/devices/perf_device.h
@@ -1,7 +1,7 @@
 #ifndef HEADER_DEVICES_PERF_DEVICE_H
 #define HEADER_DEVICES_PERF_DEVICE_H
 
-#include "libdevs.h"
+#include <f3/libdevs.h>	/* For struct device.	*/
 #include <stdint.h>	/* For type uint64_t.	*/
 
 /* Create a performance‐measuring wrapper around a device. */
