+++ include/f3/platform/block_device.h
@@ -1,7 +1,7 @@
-#ifndef HEADER_DEVICES_BLOCK_DEVICE_H
-#define HEADER_DEVICES_BLOCK_DEVICE_H
+#ifndef HEADER_PLATFORM_BLOCK_DEVICE_H
+#define HEADER_PLATFORM_BLOCK_DEVICE_H
 
-#include "libdevs.h"
+#include <f3/libdevs.h>	/* For device, reset_type.	*/
 
 /* Create a raw block device wrapper.
  * filename: path to the block device (e.g., /dev/sdx).
@@ -10,4 +10,4 @@
  */
  struct device *create_block_device(const char *filename, enum reset_type rt);
 
-#endif	/* HEADER_DEVICES_BLOCK_DEVICE_H */
+#endif	/* HEADER_PLATFORM_BLOCK_DEVICE_H */
