+++ src/f3-extra/f3fix.c
@@ -5,9 +5,9 @@
 #include <errno.h>
 #include <argp.h>
 
-#include "version.h"
-#include "libutils.h"
-#include "devices/partition.h"
+#include <f3/version.h>
+#include <f3/libutils.h>
+#include <f3/platform/partition.h>
 
 /* Sentinel for unset last sector. */
 #define SEC_UNSET ((uint64_t)UINT64_MAX)
