+++ src/f3/f3write.c
@@ -14,9 +14,9 @@
 #include <err.h>
 #include <argp.h>
 
-#include "utils.h"
-#include "libflow.h"
-#include "version.h"
+#include <f3/utils.h>
+#include <f3/libflow.h>
+#include <f3/version.h>
 
 /* Argp's global variables. */
 const char *argp_program_version = "F3 Write " F3_STR_VERSION;
