+++ src/platform/darwin/block_device.c
@@ -11,11 +11,11 @@
 #include <diskarbitration/diskarbitration.h>
 #include <CoreFoundation/CoreFoundation.h>
 
-#include "devices/block_device.h"
-#include "devices/usb_reset.h"
-#include "libutils.h"
+#include <f3/platform/block_device.h>
+#include <f3/platform/usb_reset.h>
+#include <f3/libutils.h>
 
-#include "block_device_private.h"
+#include "../private/block_device_private.h"
 
 /* XXX This is borrowing from glibc.
  * A better solution would be to return proper errors,
