+++ src/platform/linux/block_device.c
@@ -12,12 +12,11 @@
 #include <sys/stat.h>
 #include <libudev.h>
 
-#include "devices/block_device.h"
-#include "devices/usb_reset.h"
-#include "libutils.h"
+#include <f3/platform/block_device.h>
+#include <f3/libutils.h>
 
-#include "block_device_private.h"
-#include "private/private.h"
+#include "../private/block_device_private.h"
+#include "private/linux_private.h"
 
 /* XXX This is borrowing from glibc.
  * A better solution would be to return proper errors,
